Improve cmd.js and script.js conventions
Remove the main function in favor of having the module be main.  This
seems more in line with current conventions and is a bit more terse for
users and tests of the module.

Return content for stdout and stderr from the module function.  This
makes testing easier and allows callers to output whatever/however is
most appropriate.

Use commander for option parsing in cmd.js.  This was a little bit
subtle to get right, so add it to the template for convenience.
Dependency not added since it is likely to be overlooked but not
forgotten.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>