Create documentation from the Mocha specs
Mocha can generate XHTML output using the `doc` reporter.  Surround it
with an appropriate header and footer which adds some JavaScript
niceties and add it to the npm doc script.

Note:  The current use of cat only works on Unix.  As soon as the
posix-cat or cash-cat package supports the - argument we can use it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>