Update the executable file conventions
Although returning stdout and stderr content worked in a nicely
functional way, it is too limiting for scripts which have any sort of
command-line interaction, even as simple as conditionally colorizing
their output.  It also made interactivity impossible.

Instead, pass the input, output, and error streams into the function,
and only return the error code (and an Error object if returning
unexpectedly) from the function.  This still allows the calling function
to determine where output should go (and intercept/modify it) without
limiting the function's ability to interact with its input/output
sources.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>