Limit editorconfig indent to known types
Not all files should be indented with two spaces.  Only enforce
indenting for the types that we really want to limit to 2 spaces.

We do still want all files to be UTF-8 with a trailing newline, unless
explicitly excluded.

Remove LF newline requirement, since other developers may prefer CF-LF
in their local copies, which is fine.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>