Further tighten args handling
Don't allow zero-length args.  Although the intent is unambiguous, it is
likely that the caller misunderstands the format of args.  Remove this
special-case and require that if args is specified it has at least 2
elements (e.g. ["node", "script"]).

Distinguish TypeError for non-object args from RangeError for bad
length.  Throwing a TypeError for the correct type was sub-optimal.
This fixes it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>