Change ESLint max-len to ignorePattern
Previously I thought ignoreStrings et al. ignored lines which consisted
of a single string (or RegExp literal, template, or URL).  Turns out it
ignores lines which contain any of these (along with anything else).
This is not what I wanted.

Instead, write a pattern for use in ignorePattern which ignores only
what I wanted to ignore.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>