Downgrade nyc to ^11.0.0
nyc 12 requires Node 6.  On Node 4 it causes the following error:

    Transformation error for /home/travis/build/kevinoid/git-branch-is/bin/git-branch-is.js ; return original code
    Unexpected token ...
    Transformation error for /home/travis/build/kevinoid/git-branch-is/index.js ; return original code
    Unexpected token ...

Due to use of the spread operator in the instrumentation code, which
results in lack of instrumentation and incorrect coverage info (and a
test failure due to unexpected stderr output).

Stick to 11 as long as this package supports Node v4 and disable David
and Greenkeeper warnings.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>