Ignore = for operator-linebreak
The Airbnb default of "never" is problematic in cases where the
variable+method is longer than the remaining space on the line.  I think
adding parentheses in this case just adds noise.  "before" is consistent
with other operators, but I think the alignment is wrong when more
operators are present and = is a special case where the first operand
should be farther left so the operation being assigned stands alone.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>