Require Node >= 8.3
Node 8.3 added support for the object spread operator, which is
convenient (and required by the ESLint prefer-object-spread rule enabled
by airbnb).

Note:  Since security support/fixes requires installing the latest minor
version, I assume this will not be a significant issue.  If that is not
the case (e.g. a major vendor backporting fixes to Node 8.0-8.2 I will
reconsider).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>