docs(README): clarify package and template license
This template includes information for the MIT License, since that is
the license which I most commonly use for npm packages.  However, this
template may be used to create packages of any license terms without
restriction and is therefore licensed under CC0.  Clarify the licensing
of the template without complicating the licensing of derived packages.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>