feat: Better handling for detached HEAD
If git is in a "detached HEAD" state (i.e. HEAD is not on a branch),
git symbolic-ref prints

    fatal: ref HEAD is not a symbolic ref

and exits with code 128.  This causes git-branch-is to exit with an
error (including the git error message).  This behavior is not useful.

Instead, change gitBranchIs.getBranch to return an empty string when not
on a branch.  (Since empty string is not a valid branch name, is falsey,
and is easier to compare to string/pattern command line arguments.)

With this change, `git-branch-is master` and `git-branch-is --not
master` will behave the same in a detached HEAD state as on a branch
other than master.

Fixes: #58

BREAKING CHANGE: Detached HEAD is now handled like an empty branch name
(which can be matched in the usual ways) rather than causing an error.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>