Assign to exports object directly
Replacing the initial exports object after `require` will cause the
exports to be empty required as part of a circular dependency chain.
Although this can be a poor man's circular dependency checker, the
behavior can be confusing and difficult to debug.  Update the exports
object directly to avoid this.

Note: If circular dependency checking is desired, consider
https://www.npmjs.com/package/madge or
https://www.npmjs.com/package/detect-circular-deps

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>