.github: Run workflow daily
To check for exogenous breakage (e.g. in-range dependency updates which
break something, Node.js releases which break something).

I would prefer to run a build whenever a dependency updates, to catch
issues faster and reduce unnecessary builds, but I can't find a tool to
do that (Dependabot could send a PR for package-lock.json, but it is too
noisy for me).  This is workable, until I find something better.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>